chore: split utils into compass-utils and non-compass-utils
