+++ packages/compass-preferences-model/src/index.ts
@@ -23,11 +23,8 @@ export {
   getExampleConfigFile,
 } from './global-config';
 export type { ParsedGlobalPreferencesResult } from './global-config';
-export {
-  setupPreferencesAndUser,
-  getActiveUser,
-  isAIFeatureEnabled,
-} from './utils';
+export { getActiveUser, isAIFeatureEnabled } from './utils';
+export { setupPreferencesAndUser } from './compass-utils';
 export type { User, UserStorage } from './storage';
 export type { PreferencesAccess };
 export { setupPreferences };
